task_1/requirements.txt
matplotlib==3.7.1
numpy==1.23.4
Pillow==10.0.1
prettytable==3.9.0
scikit_learn==1.2.1
torch==1.13.1
torchvision==0.14.1
